Move legacy to another folder Add more sanitizing Remove moment.js calls Fix title parsing
