+++ src/plugins/tbc-ge/legacy/api.ts
@@ -1,32 +1,43 @@
 import {
-  APP_VERSION,
-  AuthV2,
-  DeviceData,
-  DeviceInfo,
-  LoginResponse,
-  PASSCODE,
+  Auth,
+  ConvertedProduct,
+  Device,
+  FetchedAccount,
+  FetchedAccounts,
+  OtpDevice,
   Preferences,
-  CardProductV2,
-  SessionV2, CardsAndAccounts, TransactionsByDateV2, FetchHistoryV2Data
+  Session
 } from './models'
 import {
-  fetchApiDepositsV2,
-  fetchApiLoansV2,
-  fetchCardAndAccountsDashboardV2,
-  fetchCardsListV2,
-  fetchCertifyLoginBySmsV2,
-  fetchConfirmTrustedDeviceV2,
-  fetchGetSessionIdV2,
-  fetchHistoryV2,
-  fetchLoginByPasscodeV2,
-  fetchLoginByPasswordV2,
-  fetchRegisterDeviceV2,
-  fetchTrustDeviceV2, fetchUnTrustDeviceV2
+  fetchAccountsList,
+  fetchCertifyLoginByPasscode,
+  fetchCertifyLoginBySms,
+  fetchConfirmTrustedDevice,
+  fetchDashboard,
+  fetchDepositDetails,
+  fetchDeposits,
+  fetchDepositStatements,
+  fetchGetLoginSalt,
+  fetchGetRequestSalt,
+  fetchHistory,
+  fetchInitHeaders,
+  fetchInitTrustedDevice,
+  fetchLoans,
+  fetchLoginByPasscode,
+  fetchLoginByPassword,
+  fetchRegisterDevice
 } from './fetchApi'
-import { InvalidOtpCodeError } from '../../errors'
+import { generateRandomString } from '../../../common/utils'
+import { InvalidOtpCodeError } from '../../../errors'
+import { getNumber, getString } from '../../../types/get'
 
-async function askOtpCodeV2 (prompt: string): Promise<string> {
-  const sms = await ZenMoney.readLine(prompt,
+async function askOtpCode (smsPrefix: string, mode: OtpDevice): Promise<string> {
+  const deviceMap: Record<OtpDevice, string> = {
+    SMS_OTP: `${smsPrefix}text message (SMS)`,
+    TOKEN_GEMALTO: 'TBC Pass App',
+    TOKEN_VASCO: 'VASCO token'
+  }
+  const sms = await ZenMoney.readLine(`Enter verification code from ${deviceMap[mode]}`,
     { inputType: 'number' })
   if (sms == null) {
     throw new InvalidOtpCodeError()
@@ -34,17 +45,23 @@ async function askOtpCodeV2 (prompt: string): Promise<string> {
   return sms.trim()
 }
 
-function generateDeviceInfo (): DeviceInfo {
-  // replace all . with empty string
-  const deviceId = ZenMoney.device.id.replace(/\./g, '_')
-  return new DeviceInfo(APP_VERSION, deviceId, ZenMoney.device.manufacturer, ZenMoney.device.model, `${ZenMoney.device.os.name} ${ZenMoney.device.os.version}`)
+function generateDevice (): Device {
+  return {
+    androidId: generateRandomString(16, '0123456789abcdef'),
+    device: ZenMoney.device.model.replace(/\s/g, '_').toLowerCase(), // Build.DEVICE
+    manufacturer: ZenMoney.device.manufacturer,
+    model: ZenMoney.device.model
+  }
 }
 
-function generateDeviceData (deviceInfo: DeviceInfo): DeviceData {
-  return DeviceData.fromDeviceInfo(deviceInfo, ZenMoney.device.os.name, ZenMoney.device.os.version)
+async function tempPatchClearCookies (): Promise<void> {
+  await ZenMoney.clearCookies()
+  const cookies = await ZenMoney.getCookies()
+  for (const cookie of cookies) {
+    await ZenMoney.setCookie(cookie.domain, cookie.name, '')
+  }
 }
-
-export async function loginV2 ({ login, password }: Preferences, auth?: AuthV2): Promise<SessionV2> {
+export async function login ({ login, password }: Preferences, auth?: Auth): Promise<Session> {
   if (ZenMoney.trustCertificates != null && typeof (ZenMoney.trustCertificates) !== 'undefined') {
     ZenMoney.trustCertificates([
       `-----BEGIN CERTIFICATE-----
@@ -125,83 +142,80 @@ WxdnLbK6zKx6+4WL9qWhGu6R+7HNPAaKOb7KXEwjV2ekr6FVZneKRFe/XivMk66O
 -----END CERTIFICATE-----`
     ])
   }
-  let session: SessionV2
-  let cookies
-  let deviceTrusted = false
-  let loginInfo: LoginResponse
-  const deviceInfo = generateDeviceInfo()
-  const deviceData = generateDeviceData(deviceInfo)
+
+  let session: Session
+  const requestSalt = await fetchGetRequestSalt()
   if (auth == null) {
-    loginInfo = await fetchLoginByPasswordV2({ username: login, password, deviceInfo, deviceData })
-    if (loginInfo.secondPhaseRequired) {
-      cookies = await fetchCertifyLoginBySmsV2(await askOtpCodeV2('Enter the code from SMS'), loginInfo.transactionId)
-    } else {
-      deviceTrusted = true
-      cookies = loginInfo.cookies
-    }
-    const registrationId = await fetchRegisterDeviceV2({ deviceName: deviceInfo.manufacturer, passcode: PASSCODE, deviceId: deviceInfo.deviceId })
-    session = {
-      cookies,
-      auth: {
-        username: login,
-        passcode: PASSCODE,
-        registrationId
-      }
-    }
+    const device = generateDevice()
+    const loginSaltInfo = await fetchGetLoginSalt(login)
+    const loginInfo = await fetchLoginByPassword({ login, password },
+      {
+        loginSalt: loginSaltInfo.salt,
+        loginHashMethod: loginSaltInfo.hashMethod,
+        requestSalt
+      }, device)
+    const ibsAccessToken = await fetchCertifyLoginBySms(await askOtpCode('', loginInfo.otpDevice), loginInfo)
+    const lightSession = { auth: { device, passcode: generateRandomString(5, '0123456789') }, ibsAccessToken }
+    const registrationId = await fetchRegisterDevice(lightSession.auth)
+    // /mbs-json/remoting/ endpoints require a JSESSIONID from login, others require it from initHeaders
+    await tempPatchClearCookies()
+    await fetchInitHeaders(lightSession)
+    const tmpSession = { auth: { ...lightSession.auth, registrationId }, ibsAccessToken }
+
+    const trustedInitInfo = await fetchInitTrustedDevice(tmpSession)
+    const trustedRegistrationId = await fetchConfirmTrustedDevice(await askOtpCode('the second ', trustedInitInfo.otpDevice), trustedInitInfo, tmpSession)
+    session = { auth: { ...tmpSession.auth, trustedRegistrationId }, ibsAccessToken }
   } else {
-    loginInfo = await fetchLoginByPasscodeV2(auth, deviceInfo, deviceData)
-    if (loginInfo.secondPhaseRequired) {
-      cookies = await fetchCertifyLoginBySmsV2(await askOtpCodeV2('Enter the code from SMS'), loginInfo.transactionId)
-    } else {
-      deviceTrusted = true
-      cookies = loginInfo.cookies
-    }
-    session = {
-      cookies,
-      auth
-    }
+    const loginInfo = await fetchLoginByPasscode(requestSalt, auth)
+    const ibsAccessToken = await fetchCertifyLoginByPasscode(loginInfo, auth)
+    session = { auth, ibsAccessToken }
+    await tempPatchClearCookies()
+    await fetchInitHeaders(session)
   }
-  const sessionId = await fetchGetSessionIdV2(cookies)
-  if (!deviceTrusted) {
-    let orderId = await fetchTrustDeviceV2(deviceData, sessionId, cookies)
-    let code = await askOtpCodeV2('Enter the second code from SMS for trusting the device')
-    let trustId: string | null = null
-
-    trustId = await fetchConfirmTrustedDeviceV2(code, orderId, cookies)
-    if (trustId == null) {
-      await ZenMoney.alert('Device was already trusted, but we don\'t have the trustedDeviceId. We need to untrust the device and trust it again')
-      await fetchUnTrustDeviceV2(deviceData, sessionId, cookies)
-      orderId = await fetchTrustDeviceV2(deviceData, sessionId, cookies)
-      code = await askOtpCodeV2('Enter the last code from SMS for trusting the device again')
-      trustId = await fetchConfirmTrustedDeviceV2(code, orderId, cookies)
-    }
-
-    if (trustId == null) {
-      throw new InvalidOtpCodeError('Device trust failed')
-    }
-
-    session.auth.trustedDeviceId = trustId
-  }
-
   return session
 }
 
-export async function fetchDepositsV2 (session: SessionV2): Promise<unknown> {
-  return await fetchApiDepositsV2(session)
+export async function fetchAccounts (session: Session): Promise<FetchedAccounts> {
+  const accounts = await fetchAccountsList(session)
+  const deposits = await fetchDeposits(session)
+  const loans = await fetchLoans(session)
+
+  return {
+    ...await fetchDashboard(session),
+    accounts: [
+      ...await Promise.all(accounts.map(async (account): Promise<FetchedAccount> => {
+        const accountType = getString(account, 'accountMatrixCategorisations[0]')
+        if (accountType === 'DEPOSITS') {
+          const externalAccountId = parseInt(getString(account, 'externalAccountId'))
+          const depositProduct = deposits.find(x => getNumber(x, 'externalAccountId') === externalAccountId)
+          assert(depositProduct != null, 'cant find deposit product', account, deposits)
+          const depositId = getNumber(depositProduct, 'id')
+          const details = await fetchDepositDetails(depositId, session)
+          return { tag: 'deposit', product: account, depositProduct, details }
+        } else {
+          return { tag: 'account', product: account }
+        }
+      })),
+      ...loans.map((x): { tag: 'loan', product: unknown } => {
+        return { tag: 'loan', product: x }
+      })
+    ]
+  }
 }
 
-export async function fetchLoansV2 (session: SessionV2): Promise<unknown> {
-  return await fetchApiLoansV2(session)
-}
-
-export async function fetchTransactionsV2 (session: SessionV2, fromDate: Date, data: FetchHistoryV2Data): Promise<TransactionsByDateV2[]> {
-  return await fetchHistoryV2(session, fromDate, data)
-}
-
-export async function fetchCardsV2 (session: SessionV2): Promise<CardProductV2[]> {
-  return await fetchCardsListV2(session)
-}
-
-export async function fetchAccountsV2 (session: SessionV2): Promise<CardsAndAccounts> {
-  return await fetchCardAndAccountsDashboardV2(session)
+export async function fetchTransactions (
+  product: ConvertedProduct,
+  fromDate: Date,
+  toDate: Date,
+  session: Session
+): Promise<unknown[]> {
+  switch (product.tag) {
+    case 'card':
+    case 'account':
+      return await fetchHistory(product.coreAccountId, session, fromDate, toDate)
+    case 'deposit':
+      return await fetchDepositStatements(product.depositId, session)
+    case 'loan':
+      return []
+  }
 }
